style: rename hyper-sync files to match camelCase convention
